refactor: Move existing rpcv6 adapters to dedicated file
