+++ rpc/v6/class.go
@@ -1,16 +1,10 @@
 package rpcv6
 
 import (
-	"encoding/json"
-	"errors"
-
-	"github.com/NethermindEth/juno/adapters/sn2core"
 	"github.com/NethermindEth/juno/core"
 	"github.com/NethermindEth/juno/core/felt"
 	"github.com/NethermindEth/juno/jsonrpc"
 	rpccore "github.com/NethermindEth/juno/rpc/rpccore"
-	"github.com/NethermindEth/juno/starknet"
-	"github.com/NethermindEth/juno/starknet/compiler"
 	"github.com/NethermindEth/juno/utils"
 )
 
@@ -40,34 +34,6 @@ type FunctionCall struct {
 	ContractAddress    felt.Felt   `json:"contract_address"`
 	EntryPointSelector felt.Felt   `json:"entry_point_selector"`
 	Calldata           []felt.Felt `json:"calldata"`
-}
-
-func adaptDeclaredClass(declaredClass json.RawMessage) (core.Class, error) {
-	var feederClass starknet.ClassDefinition
-	err := json.Unmarshal(declaredClass, &feederClass)
-	if err != nil {
-		return nil, err
-	}
-
-	switch {
-	case feederClass.V1 != nil:
-		compiledClass, cErr := compiler.Compile(feederClass.V1)
-		if cErr != nil {
-			return nil, cErr
-		}
-		return sn2core.AdaptCairo1Class(feederClass.V1, compiledClass)
-	case feederClass.V0 != nil:
-		// strip the quotes
-		base64Program := string(feederClass.V0.Program[1 : len(feederClass.V0.Program)-1])
-		feederClass.V0.Program, err = utils.Gzip64Decode(base64Program)
-		if err != nil {
-			return nil, err
-		}
-
-		return sn2core.AdaptCairo0Class(feederClass.V0)
-	default:
-		return nil, errors.New("empty class")
-	}
 }
 
 /****************************************************
